+++ entry/node/main.ts
@@ -1,1 +1,2 @@
 export * from "!/src/core/CTraderConnection";
+export * from "!/src/core/CTraderConnectionParameters";